config/generation/requirements.txt
# requirements.txt
PyYAML>=5.4
Jinja2>=3.0